perf: Prevent multiple similar events from happening in parallel
There are many ways for the "diagnostics system" to be initialized (for example, changing the contents of `package.json`). And it can run the diagnostic system multiple times, even while some other one is still running.

The `lazyCallback()` utility allows that only a single instance of this event to be executed. And when another event of the same nature occurs, it is postponed until the main event is fulfilled. Still, if many events happen at the same time, only the last one is resolved and all previous ones are just ignored.