refactor: Creating a specific Diagnostic for quick-fixes
Currently, the `Diagnostic[]` for quick-fixes are generics, and rely on `Diagnostic.code` to assign the correct quick-fix. However, the vscode displays this code in the diagnostic message.

Instead, by creating a more specific `Diagnostic` class, we can maintain the `Diagnostic.code` using the `document` as the `{target}` and the `{value}` as the own package name. And then all the specificity of the action on the package selected is done by the `PackageRelatedDiagnostic`.