fix: Avoid closing output channel after running package manager
The initial idea was that the output channel was opened to show the process and it closed automatically if everything went well, with a toast notification message.

However, as vscode does not inform if the window was already open, this function is not very acceptable. The idea was to keep it open if it was already open. In this case, an "OK" process will always close it.