feat: Major update protection
Major bump updates can be a problem. If all packages are updated at the same time, it may not be clear that some of them have major bumps and that this can lead to breaking-changes.

Now, by default, updates will be suggested "in steps". First there will be a suggestion for the latest version according to the user's current range (avoiding major bump), and only then will a new suggestion be made for a major bump, if there is a possibility.

For example, if `{ "package": "^5.0" }` and have two versions available, such as `5.1` and `6.0`, it will suggest upgrading to `5.1` first, and only the suggestion is accepted will have a new suggestion for `6.0`.

This is a configurable option, enabled by default.