fix: Check if document closed is the `package.json`
Currently, when closing any document, the diagnostics are lost. So now the package will check if the related document was actually closed.

In practice, however, this loss is recovered quickly, as diagnostics are immediately reloaded from cache. In any case, it makes no sense to lose them because of other closed documents.