feat: Limitation of parallel processes
This feature helps to reduce the amount of `npm view` processes executed during a scan.

Before, in a project with 50 packages, 50 processes were started at the same time, consuming CPU, memory and network at the same time. It is now limited to 10 concurrent processes. This increases analysis time, but ensures that it won't BOOM people's computers.

Also, this option is configurable.