refactor: Improved rendering of decorations (avoiding "blinking")
Prior to this, decorations were "blinking" whenever the document was updated. It was very-very ugly. Also, I think this was the commit I spent the most time thinking about. I must have spent about 8 hours to solve this problem!

The (big) problem: whenever we modify the document, the verification process restarts, marking the packages as "checking for update". This ensures that removed or modified lines have their decorations removed as well. Since most packages are already cached due to a previous process, the "available version" decoration comes back quickly. However, this "quickly event" takes about 15~25ms, and it made the editor to "blink" every time during any text change (displaying "checking update" and "update available" fastly).

The solution: was to delay the decorations rendering process by at least 100ms. So we make sure that we give enough time for the cache to act and everything stays as it is avoid blinking.

I noticed that the same thing happens with the diagnostics, so I'll take a look at those again.