feat: Support for pre-release updates
If there is an user-defined version indicating a pre-release version, if this is not the latest pre-release available then it will be suggested.

Eg.: `^13.0.7-canary.3`, while there is a new pre-release `13.0.7-canary.6`.

Important: if the user informs a pre-release on his own, it is understood that he wants to remain in a pre-release.